requirements.txt
Django==2.1.2
django-filter==2.0.0
django-request-logging==0.6.6
djangorestframework==3.9.0
Markdown==3.0.1